+++ apps/docs/components/Providers.tsx
@@ -1,13 +1,13 @@
 'use client'
 
-import { ModalRoot } from '@woon-ui/dialog'
+import { DialogRuntime } from '@woon-ui/dialog'
 import { Toaster } from '@woon-ui/toast'
 
 export function Providers({ children }: { children: React.ReactNode }) {
   return (
     <>
       {children}
-      <ModalRoot />
+      <DialogRuntime />
       <Toaster position="bottom-right" />
     </>
   )