+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # docs
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @woon-ui/dialog@0.4.0
+  - @woon-ui/react@0.4.0
+  - @woon-ui/drawer@0.2.4
+
 ## 0.1.5
 
 ### Patch Changes